+++ CJBCheatsMenu/release-notes.md
@@ -3,6 +3,7 @@
 # Release notes
 ## Upcoming release
 * Added 'instant cooldowns' cheat which removes tool/weapon cooldowns.
+* Merged 'increase movement speed' and 'move speed' options.
 * Fixed instantly-grown wild seeds different from naturally-grown wild seeds (e.g. not applying botanist bonus, not randomized, etc).
 * Fixed instantly-grown crops not generating giant crops.
 * Improved translations. Thanks to Akir0 (updated French), Annosz (added Hungarian), and D0n-A (updated Russian)!